requirements.txt
git+https://github.com/Project-Platypus/Platypus.git
https://github.com/s4w3d0ff/python-poloniex/archive/v0.4.5.zip
blist===1.3.6
numpy==1.13
python-telegram-handler==2.0.1
pymongo==3.5.1
TA-Lib==0.4.14
rbfopt==4.0.0.post2
